requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# DB — MongoDB 
motor==3.4.0
pymongo[srv]==4.7.2

# ML
scikit-learn==1.5.0
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.3
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

# MLflow
mlflow>=2.14.0

# Cache / messaging
redis==5.0.6
aioredis==2.0.1

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Monitoring
prometheus-fastapi-instrumentator==7.0.0
structlog==24.2.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0

# Utils
pydantic==2.7.4
pydantic-settings==2.3.3
tenacity==8.3.0
setuptools
dill
